day-5

+++ MoonValleyNursery/src/main/java/com/MVNursery/controller/PlantController.java
@@ -14,9 +14,14 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
+import com.MVNursery.exception.SessionException;
 import com.MVNursery.model.Plant;
+import com.MVNursery.model.Session;
+import com.MVNursery.model.UserType;
+import com.MVNursery.service.ISessionService;
 import com.MVNursery.service.PlantService;
 
 @RestController
@@ -26,22 +31,50 @@ public class PlantController {
 	@Autowired
 	private PlantService pService;
 	
-	@PostMapping(value="")
-	public ResponseEntity<Plant> addPlantHandler(@Valid @RequestBody Plant plant){
+	@Autowired
+	private ISessionService sessionService;
+	
+	@PostMapping(value="/{adminId}/{sessionKey}")
+	public ResponseEntity<Plant> addPlantHandler(@Valid @RequestBody Plant plant,@PathVariable("adminId") Integer adminId,
+			@PathVariable("sessionKey") String sessionKey)throws SessionException{
+		
+		Session session = sessionService.getASessionByKey(sessionKey);
+		if (session.getUserId() == adminId && session.getUserType() == UserType.ADMIN) {
+
 		Plant savePlant = pService.addPlant(plant);
 		return new ResponseEntity<Plant>(savePlant, HttpStatus.CREATED);
+		}else {
+			throw new SessionException("Please login with the correct credentials");
+		}
 	}
 	
-	@PutMapping(value="")
-	public ResponseEntity<Plant> updatePlantHandler(@Valid @RequestBody Plant plant){
+	@PutMapping(value="/{adminId}/{sessionKey}")
+	public ResponseEntity<Plant> updatePlantHandler(@Valid @RequestBody Plant plant,@PathVariable("adminId") Integer adminId,
+			@PathVariable("sessionKey") String sessionKey)throws SessionException{
+		
+		Session session = sessionService.getASessionByKey(sessionKey);
+		if (session.getUserId() == adminId && session.getUserType() == UserType.ADMIN) {
+
+		
 		Plant updatePlant = pService.addPlant(plant);
 		return new ResponseEntity<Plant>(updatePlant, HttpStatus.ACCEPTED);
+		}else {
+			throw new SessionException("Please login with the correct credentials");
+		}
 	}
 	
 	@DeleteMapping(value="/{id}")
-	public ResponseEntity<Plant> deletePlantHandler(@PathVariable Integer id){
+	public ResponseEntity<Plant> deletePlantHandler(@PathVariable Integer id,@RequestParam("sessionKey") String sessionKey,
+			@RequestParam("adminId") Integer adminId)throws SessionException{
+		
+		Session session = sessionService.getASessionByKey(sessionKey);
+		if (session.getUserId() == adminId && session.getUserType() == UserType.ADMIN) {
+
 		Plant deletePlant = pService.deletePlant(id);
 		return new ResponseEntity<Plant>(deletePlant,HttpStatus.OK);
+		}else {
+			throw new SessionException("Please login with the correct credentials");
+		}
 	}
 	
 	@GetMapping(value="/{id}")
